rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "adamvalik",email = "xvalik05@stud.fit.vutbr.cz"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "weaviate-client (>=4.13.2,<5.0.0)",
    "unstructured (>=0.17.2,<0.18.0)",
    "onnxruntime (==1.21.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-api-python-client (>=2.166.0,<3.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "openai (>=1.72.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "unstructured-inference (>=0.8.10,<0.9.0)",
    "pdfminer-six (>=20250327,<20250328)",
    "pi-heif (>=0.22.0,<0.23.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "markdown (>=3.7,<4.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "unstructured-pytesseract (>=0.3.15,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "ragas (>=0.2.14,<0.3.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
